- latitude and longitude should be double and not int/long - changed it back to "populate" every time with new values

+++ Rove/src/com/fourkins/rove/sqlite/PostsDataSource.java
@@ -67,8 +67,8 @@ public class PostsDataSource {
 
         post.setId(cursor.getLong(0));
         post.setUserName(cursor.getString(1));
-        post.setLatitude(cursor.getLong(2));
-        post.setLongitude(cursor.getLong(3));
+        post.setLatitude(cursor.getDouble(2));
+        post.setLongitude(cursor.getDouble(3));
         post.setMessage(cursor.getString(4));
 
         return post;
